Move statetransfer to peer package
